requirements.txt
oci
python-dotenv